requirements.txt
python-consul==1.1.0
Flask==3.0.3
flask_jwt_extended==4.6.0
flask_limiter==3.8.0
mysql-connector-python==8.0.33
pika==1.3.2
pybreaker==1.2.0
pytest==8.3.3
redis==5.0.8
Requests==2.32.3
pytest-mock==3.6.1